progress
I tried to make the role management function but I have some errors that I will be fixing later on. I have created a register and login page with error handlings and added a forget password. I have to fix the role management to make the favourites page function. I did the 'Byedit' products page and did a product details page with reviews and add to cart button. I included rating stars and a review linked product. I did the cart as a side cart which functions.