progress
In the account page I removed 'buy again' and added a sign out button instead, as I already have 'my orders' and I can include 'buy again' in that section. I added that the user can add a review and a comment to the review in the product details page. I created the pages of the naavbar dropdown. I finished the face care and by edit pages. the user can add the products to cart and view each product with details. I made the layout of the side cart more aesthetic. I started to work on the checkout process.