progress
For the register , I added 'if you already have an account log in' and added the database to the login page swell. by doing that I made the add to favourites products function. but I still have bugs as the products arent showing. I added a forget password to the login but I still have to work on it.